+++ R/XenaQueryProbeMap.R
@@ -24,7 +24,7 @@ XenaQueryProbeMap <- function(x) {
   } else {
     query <- query %>%
       dplyr::rename(hosts = XenaHosts, datasets = ProbeMap) %>%
-      dplyr::mutate(url = ifelse(.data$XenaHostNames %in% c("gdcHub", "gdcHubV18"),
+      dplyr::mutate(url = ifelse(.data$XenaHostNames == "gdcHub",
         file.path(hosts, "download", url_encode(basename(datasets))),
         file.path(hosts, "download", url_encode(datasets))
       )) %>%